Components set - remove unnecessary CSS
Taking a new approach to the components with styled-components.

This will make all the components re-usable for next projects but also able to be customised.

Can still use styles from Material - just delete the css from the component and is clean to accept Material styles.

Will be aiming for components to be custom with icons imported from Material only